update to hashtag regexp...
Well, as I see, the unicode problem is bigger than I've thought. Twitter only uses a subset of unicode alphabet letters (which contain more than a-zA-Z1-9... etc.). This code is directly from twitter. But i'm not sure if you can use it or not. Because this library is MIT licenced, twitter code has Apache 2 licence. I don't get the licence things well...

https://github.com/twitter/twitter-text/blob/master/js/twitter-text.js

And here is a regexp range generator. (click on some checkboxes and you'll see what it does :) )

http://apps.timwhitlock.info/js/regex

I don't know if it helps. But this is all for now.