Support returning reactElement from replaceFn
This meant being able to pass objects around with turning them into a string. I added a doJoin option which makes function return an array of nodes, rather than joining them together to a single string.

It didn't make sense to add React as a dependency, so I let the user pass it in themselves.